samples/functions-pep/requirements.txt
azure-functions
requests
pyjwt
